src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --all-extras --no-hashes --no-dev
aiolimiter==1.2.1
annotated-types==0.7.0
anyio==4.8.0
apscheduler==3.11.0
beautifulsoup4==4.9.3
blinker==1.9.0
cachetools==5.5.0
cchardet==2.1.7
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
colorama==0.4.6 ; sys_platform == 'win32'
dawg-python==0.7.2
decorator==5.1.1
deprecation==2.1.0
docopt==0.6.2
exceptiongroup==1.2.2 ; python_full_version < '3.11'
flask==2.3.3
functions-framework==3.8.2
geographiclib==1.52
geopy==2.2.0
google-api-core==2.24.0
google-auth==2.19.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-logging==3.5.0
google-cloud-pubsub==2.16.0
google-cloud-secret-manager==2.16.0
google-cloud-storage==2.10.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.62.3
gunicorn==23.0.0 ; sys_platform != 'win32'
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.7
intervaltree==3.1.0
ipymarkup==0.9.0
itsdangerous==2.2.0
jinja2==3.1.5
lxml==5.4.0
markupsafe==3.0.2
multidict==6.2.0
natasha==1.4.0
navec==0.10.0
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
propcache==0.3.1
proto-plus==1.25.0
protobuf==4.25.5
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.5
pydantic-core==2.27.2
pydantic-settings==2.7.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==21.11.1
pytz==2021.1
razdel==0.5.0
requests==2.32.3
retry2==0.9.5
rsa==4.9
six==1.17.0
slovnet==0.6.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlalchemy==1.4.54
typing-extensions==4.12.2
tzdata==2024.2 ; sys_platform == 'win32'
tzlocal==5.2
urllib3==1.26.19
watchdog==6.0.0
werkzeug==3.1.3
yandex-geocoder==3.0.1
yargy==0.16.0
yarl==1.18.3
